backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt
pydantic>=2.0
pydantic-settings
python-dotenv
python-multipart
asyncpg
